104-fibonacci

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,9 +9,9 @@
 int main(void)
 {
 	int count;
-	unsigned long fib1 = 0, fib2 = 1, sum;
-	unsigned long fib1_half1, fib1_half2, fib2_half1, fib2_half2;
-	unsigned long half1, half2;
+	int fib1 = 0, fib2 = 1, sum;
+	int fib1_half1, fib1_half2, fib2_half1, fib2_half2;
+	int half1, half2;
 
 	for (count = 0; count < 92; count++)
 	{
